+++ athenacli/main.py
@@ -62,7 +62,9 @@ class AthenaCli(object):
     def __init__(self, region, aws_access_key_id, aws_secret_access_key,
                  s3_staging_dir, athenaclirc, profile, database):
 
-        config_files = (DEFAULT_CONFIG_FILE, athenaclirc)
+        config_files = [DEFAULT_CONFIG_FILE]
+        if os.path.exists(athenaclirc):
+            config_files.append(athenaclirc)
         _cfg = self.config = read_config_files(config_files)
 
         self.init_logging(_cfg['main']['log_file'], _cfg['main']['log_level'])
@@ -523,7 +525,7 @@ For more details about the error, you can check the log file: %s''' % (ATHENACLI
                 editing_mode = EditingMode.VI
             else:
                 editing_mode = EditingMode.EMACS
-            
+
             self.prompt_app = PromptSession(
                 lexer=PygmentsLexer(Lexer),
                 reserve_space_for_menu=self.get_reserved_space(),
@@ -621,7 +623,7 @@ def cli(execute, region, aws_access_key_id, aws_secret_access_key,
       - athenacli
       - athenacli my_database
     '''
-    if (athenaclirc == ATHENACLIRC) and (not os.path.exists(os.path.expanduser(ATHENACLIRC))):
+    if athenaclirc and (athenaclirc == ATHENACLIRC) and (not os.path.exists(os.path.expanduser(ATHENACLIRC))):
         err_msg = '''
         Welcome to athenacli!